+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 LATEST CHANGES
 ==============
 
+2023-01-19
+----------
+- GPS fusion integration
+  - New param `pos_tracking.gps_fusion` to enable GPS fusion
+  - New param `pos_tracking.gps_fix_topic` name of the topic containing GPS Fix data of type `sensor_msgs/NavSatFix`
+  - Add `nmea_msgs` dependency
+  - Add GPS Fix Diagnostic
+
 2023-01-16
 ----------
 - Removed annoying build log messages. Only warning regarding unsupported ROS2 distributions will be displayed when required.